[BUGFIX] existing date values were being cleared
Problem:

When a user edits a resource
with a `Date`, `Time`, or `DateTime` attribute,
and the attribute already has a value,
the date-time picker clears out the value.

Unless the user re-fills the field,
the value will be reset and will cause an error
when they hit "submit".

Solution:

Update the `datetimepicker` date format
to match the Rails default.